Handle the case if there is no SearchWidgetProvider available:
If either no gapps package is installed,
or one of the minimal/core gapps packages is installed,
possibly there is no SearchWidgetProvider available.

If there is no SearchWidgetProvider available;
- Fix FC
- Remove the show/hide search bar preference

The default value for show/hide the search bar is show,
I don`t want to change it to hide,
in most cases, the SearchWidgetProvider is available,
if it doesn`t set the (non default) value to hide.